move functions to utils folder
