+++ packages/nextjs/utils/example-zk/copy-to-clipboard.ts
@@ -1,0 +1,28 @@
+export function copyToClipboard({
+  text,
+  successCallback,
+  errorCallback,
+}: {
+  text: string;
+  successCallback?: () => void;
+  errorCallback?: () => void;
+}) {
+  if (navigator) {
+    navigator.clipboard.writeText(text).then(
+      function () {
+        if (successCallback) {
+          successCallback();
+        }
+      },
+      function () {
+        if (errorCallback) {
+          errorCallback();
+        }
+      },
+    );
+  } else {
+    if (errorCallback) {
+      errorCallback();
+    }
+  }
+}
