+++ packages/nextjs/utils/example-zk/short-hash-string.ts
@@ -1,0 +1,6 @@
+export function shortenHashString(hash: string) {
+  const prefix = hash.substring(0, 2);
+  const suffix = hash.substring(hash.length - 4);
+
+  return prefix + "..." + suffix;
+}
